⚡️ 라우터 변경
- AuthPage 삭제
- 로그인 및 회원가입 페이지 추가
- 라우터 변경

#4
